--- src/shared/ui/button/index.ts
@@ -1,1 +1,0 @@
-export { Button } from './ui';
