--- src/shared/ui/dialog/index.ts
@@ -1,9 +1,0 @@
-export {
-  Dialog,
-  DialogTrigger,
-  DialogContent,
-  DialogPortal,
-  DialogOverlay,
-  DialogHeader,
-  DialogTitle,
-} from './ui';
