+++ src/shared/ui/index.ts
@@ -1,0 +1,33 @@
+// Button Components
+export { Button } from './button';
+
+// Card Components
+export { Card, CardHeader, CardTitle, CardContent } from './card';
+
+// Dialog Components
+export {
+  Dialog,
+  DialogTrigger,
+  DialogPortal,
+  DialogOverlay,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+} from './dialog';
+
+// Form Components
+export { Input } from './input';
+export { Textarea } from './textarea';
+
+// Select Components
+export {
+  Select,
+  SelectGroup,
+  SelectValue,
+  SelectTrigger,
+  SelectContent,
+  SelectItem,
+} from './select';
+
+// Table Components
+export { Table, TableHeader, TableBody, TableRow, TableHead, TableCell } from './table';
