--- src/shared/ui/input/index.ts
@@ -1,1 +1,0 @@
-export { Input } from './ui';
