--- src/shared/ui/table/index.ts
@@ -1,1 +1,0 @@
-export { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from './ui';
