--- src/shared/ui/textarea/index.ts
@@ -1,1 +1,0 @@
-export { Textarea } from './ui';
